# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: SchoolHub API
  version: 1.0.0
  description: Multi-tenant School Management SaaS — see the specification in docs/.
paths:
...
  /api/v1/staff-documents/{id}:
    get:
      operationId: staff_documents_retrieve
      description: Top-level access for `DELETE /staff-documents/{id}` and `:verify`.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this staff document.
        required: true
      tags:
      - staff-documents
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffDocument'
          description: ''
    delete:
      operationId: staff_documents_destroy
      description: Top-level access for `DELETE /staff-documents/{id}` and `:verify`.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this staff document.
        required: true
      tags:
      - staff-documents
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/staff-qualifications/{id}:
    get:
      operationId: staff_qualifications_retrieve
      description: Top-level access for `PATCH /staff-qualifications/{id}` and `:verify`.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this staff qualification.
        required: true
      tags:
      - staff-qualifications
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffQualification'
          description: ''
    patch:
      operationId: staff_qualifications_partial_update
      description: Top-level access for `PATCH /staff-qualifications/{id}` and `:verify`.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this staff qualification.
        required: true
      tags:
      - staff-qualifications
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedStaffQualification'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedStaffQualification'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedStaffQualification'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffQualification'
          description: ''
  /api/v1/staff/{id}:
    get:
      operationId: staff_retrieve
      description: Staff master records (module doc §5.1).
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this staff.
        required: true
      tags:
      - staff
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
          description: ''
    patch:
      operationId: staff_partial_update
      description: Staff master records (module doc §5.1).
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this staff.
        required: true
      tags:
      - staff
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedStaff'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedStaff'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedStaff'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
          description: ''
    delete:
      operationId: staff_destroy
      description: Staff master records (module doc §5.1).
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this staff.
        required: true
      tags:
      - staff
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    EmploymentStatusEnum:
      enum:
      - active
      - on_leave
      - suspended
      - resigned
      - retired
      - terminated
      type: string
      description: |-
        * `active` - Active
        * `on_leave` - On leave
        * `suspended` - Suspended
        * `resigned` - Resigned
        * `retired` - Retired
        * `terminated` - Terminated
    EmploymentTypeEnum:
      enum:
      - full_time
      - part_time
      - contract
      - visiting
      type: string
      description: |-
        * `full_time` - Full time
        * `part_time` - Part time
        * `contract` - Contract
        * `visiting` - Visiting
...
    GenderEnum:
      enum:
      - male
      - female
      - other
      - unspecified
      type: string
      description: |-
        * `male` - Male
        * `female` - Female
        * `other` - Other
        * `unspecified` - Unspecified
...
    PatchedStaff:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        employee_number:
          type: string
          maxLength: 32
        user_id:
          type: string
          format: uuid
          nullable: true
          description: users(id) — portal account, tenant-checked at write time.
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        gender:
          $ref: '#/components/schemas/GenderEnum'
        date_of_birth:
          type: string
          format: date
          nullable: true
        photo_file_id:
          type: string
          format: uuid
          nullable: true
        staff_type:
          $ref: '#/components/schemas/StaffTypeEnum'
        campus_id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
          nullable: true
        designation_id:
          type: string
          format: uuid
          nullable: true
        reports_to_staff_id:
          type: string
          format: uuid
          nullable: true
        employment_type:
          $ref: '#/components/schemas/EmploymentTypeEnum'
        employment_status:
          allOf:
          - $ref: '#/components/schemas/EmploymentStatusEnum'
          readOnly: true
        joining_date:
          type: string
          format: date
        exit_date:
          type: string
          format: date
          readOnly: true
          nullable: true
        exit_reason:
          type: string
          readOnly: true
          nullable: true
        email:
          nullable: true
          oneOf:
          - type: string
            format: email
            maxLength: 254
          - type: string
            maxLength: 0
        phone:
          type: string
          maxLength: 32
        national_id:
          type: string
          nullable: true
          maxLength: 64
        public_bio:
          type: string
          nullable: true
          description: Opt-in website-published bio (§10).
        address:
          nullable: true
        custom_fields: {}
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    PatchedStaffQualification:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        staff_id:
          type: string
          format: uuid
          readOnly: true
        qualification_type:
          $ref: '#/components/schemas/QualificationTypeEnum'
        title:
          type: string
          maxLength: 200
        institution:
          type: string
          nullable: true
          maxLength: 200
        field_of_study:
          type: string
          nullable: true
          maxLength: 120
        year_awarded:
          type: integer
          maximum: 32767
          minimum: -32768
          nullable: true
        grade:
          type: string
          nullable: true
          maxLength: 50
        document_file_id:
          type: string
          format: uuid
          nullable: true
        verification_status:
          allOf:
          - $ref: '#/components/schemas/VerificationStatusEnum'
          readOnly: true
        verified_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: users(id).
        verified_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    QualificationTypeEnum:
      enum:
      - degree
      - diploma
      - certification
      - training
      - license
      type: string
      description: |-
        * `degree` - Degree
        * `diploma` - Diploma
        * `certification` - Certification
        * `training` - Training
        * `license` - License
...
    Staff:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        employee_number:
          type: string
          maxLength: 32
        user_id:
          type: string
          format: uuid
          nullable: true
          description: users(id) — portal account, tenant-checked at write time.
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        gender:
          $ref: '#/components/schemas/GenderEnum'
        date_of_birth:
          type: string
          format: date
          nullable: true
        photo_file_id:
          type: string
          format: uuid
          nullable: true
        staff_type:
          $ref: '#/components/schemas/StaffTypeEnum'
        campus_id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
          nullable: true
        designation_id:
          type: string
          format: uuid
          nullable: true
        reports_to_staff_id:
          type: string
          format: uuid
          nullable: true
        employment_type:
          $ref: '#/components/schemas/EmploymentTypeEnum'
        employment_status:
          allOf:
          - $ref: '#/components/schemas/EmploymentStatusEnum'
          readOnly: true
        joining_date:
          type: string
          format: date
        exit_date:
          type: string
          format: date
          readOnly: true
          nullable: true
        exit_reason:
          type: string
          readOnly: true
          nullable: true
        email:
          nullable: true
          oneOf:
          - type: string
            format: email
            maxLength: 254
          - type: string
            maxLength: 0
        phone:
          type: string
          maxLength: 32
        national_id:
          type: string
          nullable: true
          maxLength: 64
        public_bio:
          type: string
          nullable: true
          description: Opt-in website-published bio (§10).
        address:
          nullable: true
        custom_fields: {}
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - campus_id
      - created_at
      - employment_status
      - exit_date
      - exit_reason
      - first_name
      - id
      - joining_date
      - last_name
      - phone
      - staff_type
      - updated_at
    StaffDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        staff_id:
          type: string
          format: uuid
          readOnly: true
        file_id:
          type: string
          format: uuid
        document_type:
          type: string
          description: 'Tenant-extensible; seeded values: contract, national_id, resume,
            police_clearance, medical_certificate, other.'
          maxLength: 50
        title:
          type: string
          maxLength: 200
        notes:
          type: string
          nullable: true
          maxLength: 500
        verification_status:
          allOf:
          - $ref: '#/components/schemas/VerificationStatusEnum'
          readOnly: true
        verified_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: users(id).
        verified_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        expires_at:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - document_type
      - file_id
      - id
      - staff_id
      - title
      - updated_at
      - verification_status
      - verified_at
      - verified_by
...
    StaffQualification:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        staff_id:
          type: string
          format: uuid
          readOnly: true
        qualification_type:
          $ref: '#/components/schemas/QualificationTypeEnum'
        title:
          type: string
          maxLength: 200
        institution:
          type: string
          nullable: true
          maxLength: 200
        field_of_study:
          type: string
          nullable: true
          maxLength: 120
        year_awarded:
          type: integer
          maximum: 32767
          minimum: -32768
          nullable: true
        grade:
          type: string
          nullable: true
          maxLength: 50
        document_file_id:
          type: string
          format: uuid
          nullable: true
        verification_status:
          allOf:
          - $ref: '#/components/schemas/VerificationStatusEnum'
          readOnly: true
        verified_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: users(id).
        verified_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - qualification_type
      - staff_id
      - title
      - updated_at
      - verification_status
      - verified_at
      - verified_by
    StaffTypeEnum:
      enum:
      - teaching
      - non_teaching
      type: string
      description: |-
        * `teaching` - Teaching
        * `non_teaching` - Non-teaching
...
    VerificationStatusEnum:
      enum:
      - pending
      - verified
      - rejected
      type: string
      description: |-
        * `pending` - Pending
        * `verified` - Verified
        * `rejected` - Rejected
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT